feat(6750): extract kolme runtime commit http transport tests
