+++ crates/kamn-core/tests/kolme_runtime_commit_http_transport/fork_profile_contract_tests.rs
@@ -1,0 +1,6 @@
+use super::*;
+
+#[path = "fork_profile_contract_tests/profile_submit_contract_tests.rs"]
+mod profile_submit_contract_tests;
+#[path = "fork_profile_contract_tests/signed_payload_contract_tests.rs"]
+mod signed_payload_contract_tests;
