+++ crates/kamn-core/tests/kolme_runtime_commit_http_transport/http_transport_contract_tests.rs
@@ -1,0 +1,10 @@
+use super::*;
+
+#[path = "http_transport_contract_tests/validation_request_contract_tests.rs"]
+mod validation_request_contract_tests;
+#[path = "http_transport_contract_tests/submit_finality_contract_tests.rs"]
+mod submit_finality_contract_tests;
+#[path = "http_transport_contract_tests/response_parsing_contract_tests.rs"]
+mod response_parsing_contract_tests;
+#[path = "http_transport_contract_tests/connection_pool_contract_tests.rs"]
+mod connection_pool_contract_tests;
